@@ -16,7 +16,7 @@
         }
     },
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.0",
         "guzzlehttp/guzzle": "^7.3",
         "php-arsenal/salesforce-soap-client": "^2.0",
         "symfony/console": "^5.2",
